+++ src/command_modules/azure-cli-resource/azure/cli/command_modules/resource/_help.py
@@ -152,6 +152,155 @@ helps['account lock update'] = """
           text: >
             az account lock update --name lockName --notes newNotesHere --lock-type CanNotDelete
     """
+helps['account management-group'] = """
+    type: group
+    short-summary: Manage Azure Management Groups.
+"""
+
+helps['account management-group subscription'] = """
+    type: group
+    short-summary: Subscription operations for Management Groups.
+"""
+
+helps['account management-group list'] = """
+    type: command
+    short-summary: List all management groups.
+    long-summary: List of all management groups in the current tenant.
+    examples:
+        - name: List all management groups
+          text: >
+             az account management-group list
+"""
+
+helps['account management-group show'] = """
+    type: command
+    short-summary: Get a specific management group.
+    long-summary: Get the details of the management group.
+    parameters:
+        - name: --name -n
+          type: string
+          short-summary: Name of the management group.
+        - name: --expand -e
+          type: bool
+          short-summary: If given, lists the children in the first level of hierarchy.
+        - name: --recurse -r
+          type: bool
+          short-summary: If given, lists the children in all levels of hierarchy.
+    examples:
+        - name: Get a management group.
+          text: >
+             az account management-group show --name GroupName
+        - name: Get a management group with children in the first level of hierarchy.
+          text: >
+             az account management-group show --name GroupName -e
+        - name: Get a management group with children in all levels of hierarchy.
+          text: >
+             az account management-group show --name GroupName -e -r
+"""
+
+helps['account management-group create'] = """
+    type: command
+    short-summary: Create a new management group.
+    long-summary: Create a new management group.
+    parameters:
+        - name: --name -n
+          type: string
+          short-summary: Name of the management group.
+        - name: --display-name -d
+          type: string
+          short-summary: Sets the display name of the management group. If null, the group name is set as the display name.
+        - name: --parent -p
+          type: string
+          short-summary: Sets the parent of the management group. Can be the fully qualified id or the name of the management group. If null, the root tenant group is set as the parent.
+    examples:
+        - name: Create a new management group.
+          text: >
+             az account management-group create --name GroupName
+        - name: Create a new management group with a specific display name.
+          text: >
+             az account management-group create --name GroupName --display-name DisplayName
+        - name: Create a new management group with a specific parent.
+          text: >
+             az account management-group create --name GroupName --parent ParentId/ParentName
+        - name: Create a new management group with a specific display name and parent.
+          text: >
+             az account management-group create --name GroupName --display-name DisplayName --parent ParentId/ParentName
+"""
+
+helps['account management-group update'] = """
+    type: command
+    short-summary: Update an existing management group.
+    long-summary: Update an existing management group.
+    parameters:
+        - name: --name -n
+          type: string
+          short-summary: Name of the management group.
+        - name: --display-name -d
+          type: string
+          short-summary: Updates the display name of the management group. If null, no change is made.
+        - name: --parent -p
+          type: string
+          short-summary: Update the parent of the management group. Can be the fully qualified id or the name of the management group. If null, no change is made.
+    examples:
+        - name: Update an existing management group with a specific display name.
+          text: >
+             az account management-group update --name GroupName --display-name DisplayName
+        - name: Update an existing management group with a specific parent.
+          text: >
+             az account management-group update --name GroupName --parent ParentId/ParentName
+        - name: Update an existing management group with a specific display name and parent.
+          text: >
+             az account management-group update --name GroupName --display-name DisplayName --parent ParentId/ParentName
+"""
+
+helps['account management-group delete'] = """
+    type: command
+    short-summary: Delete an existing management group.
+    long-summary: Delete an existing management group.
+    parameters:
+        - name: --name -n
+          type: string
+          short-summary: Name of the management group.
+    examples:
+        - name: Delete an existing management group
+          text: >
+             az account management-group delete --name GroupName
+"""
+
+helps['account management-group subscription add'] = """
+    type: command
+    short-summary: Add a subscription to a management group.
+    long-summary: Add a subscription to a management group.
+    parameters:
+        - name: --name -n
+          type: string
+          short-summary: Name of the management group.
+        - name: --subscription -s
+          type: string
+          short-summary: Subscription Id or Name
+    examples:
+        - name: Add a subscription to a management group.
+          text: >
+             az account management-group subscription add --name GroupName --subscription Subscription
+"""
+
+helps['account management-group subscription remove'] = """
+    type: command
+    short-summary: Remove an existing subscription from a management group.
+    long-summary: Remove an existing subscription from a management group.
+    parameters:
+        - name: --name -n
+          type: string
+          short-summary: Name of the management group.
+        - name: --subscription -s
+          type: string
+          short-summary: Subscription Id or Name
+    examples:
+        - name: Remove an existing subscription from a management group.
+          text: >
+             az account management-group subscription remove --name GroupName --subscription Subscription
+"""
+
 helps['policy'] = """
     type: group
     short-summary: Manage resource policies.