+++ src/command_modules/azure-cli-acr/HISTORY.rst
@@ -2,6 +2,11 @@
 
 Release History
 ===============
+2.0.26
+++++++
+* Allow VSTS as a remote source location
+* Add 'acr import' command.
+
 2.0.25
 ++++++
 * Improve manifest delete confirmation on tags